--- src/swagger/index.ts
@@ -1,1 +1,0 @@
-export * from './build-swagger-by-services';
